# runtime/internal/openapidocs/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "type": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "approvalapi_AnswerRequest": {
        "properties": {
          "answer": {
            "type": "string"
          },
          "approved": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "runtimetypes_HITLApproval": {
        "properties": {
          "agentName": {
            "type": "string"
          },
          "argsSummary": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "diff": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "instanceId": {
            "type": "string"
          },
          "matchedRule": {
            "type": "integer"
          },
          "missionId": {
            "type": "string"
          },
          "onTimeout": {
            "type": "string"
          },
          "policyName": {
            "type": "string"
          },
          "resolution": {},
          "resolvedAt": {
            "format": "date-time",
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "toolName": {
            "type": "string"
          },
          "toolsName": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API served by `contenox serve` under /api. Generated from route annotations by tools/openapi-gen.",
    "title": "Contenox Runtime API",
    "version": "v0.36.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/approvals": {
      "get": {
        "operationId": "approval_list",
        "parameters": [
          {
            "description": "The maximum number of items to return per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/runtimetypes_HITLApproval"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "list returns pending approvals, newest first — the inbox itself.",
        "tags": [
          "approval"
        ]
      }
    },
    "/approvals/{id}": {
      "post": {
        "operationId": "approval_answer",
        "parameters": [
          {
            "description": "The unique ID of the approval.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/approvalapi_AnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "answer resolves a pending approval.",
        "tags": [
          "approval"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}